pyproject.toml
[tool.poetry]
name = "rustyaml"
version = "0.1.0"
license = "MIT"
description = "Python bindings for rust-yaml"
authors = ["Will Eaton <me@wseaton.com>"]
build = "build.py"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Rust",
    "License :: OSI Approved :: MIT License"
]


[tool.poetry.dependencies]
python = ">=3.6.0,<4.0"

[tool.poetry.dev-dependencies]
setuptools-rust = "^0.11.6"
pytest = "^6.2.1"
black = "^20.8b1"
isort = "^5.6.4"
ipython = "^7"
pytest-benchmark = {extras = ["histogram"], version = "^3.2.3"}
PyYAML = "^5.3.1"


[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=41.0.0", "wheel", "setuptools_rust>=0.10.2"]
build-backend = "poetry.core.masonry.api"
